+++ src/Contract/RouteCreateActionInterface.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\Base\Contract;
+
+use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
+
+interface RouteCreateActionInterface
+{
+    /**
+     * @return iterable<RouteCreateResult>
+     *
+     * @throws UnsupportedStorageKeyException
+     *
+     * @TODO add exception for creation failures
+     */
+    public function create(RouteCreateParams $params): iterable;
+}